Run script in their own env from master
I wanted to source all the scripts that we run on wednesday on any regular basis from master_episitrep.R

The fact that there is a rm(list = ls()) at the begining of setup.R has weird consequences.

1) It messes up with parameterised repports (since it removes list, it suppress params), but it can be dealt by detecting whether it had been ran prior to the script.

2) It messes up the master_episitrep script since it removew paths and so forth.

3) But not runing it means risking using data that had been altered in another script (e.g. filtered).

So I propose to run these scripts in their own environment from the master_script.